requirements.txt
boto3==1.40.44
botocore==1.40.44
jmespath==1.0.1
psycopg2-binary==2.9.10
py4j==0.10.9.7
pyspark==3.5.1
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
urllib3==1.26.20
